Implement new `Compiler` API
Implement `Compiler::register` and `Compiler::get_identifier` by
redesigning `SymbolTable` which is migrated into its own module.
